Add packages to fields in init_atmosphere core Registry based on output packages
In general, for fields that are written to output streams by the init_atmosphere
core, those fields only need to be allocated if they are active in any output
stream. This implies that we can apply packages from the output streams to the
fields themselves to potentially save memory (depending on which pre-processing
stages are active). Note that when fields are written to multiple output
streams, we must add all output packages to a field; for example, the 'sst' and
'xice' fields may be written to both the "output" and "surface" streams and
therefore must be associated with both the "met_stage_out" and the "sfc_update"
packages.