Add optional maximum search distance parameter to mpas_kd_search routine
The mpas_kd_search routine now accepts an optional argument, max_distance,
that specifies an upper bound on the distance from the query point to look
for nearest points in the tree. If no point in the tree is within the specified
maximum distance from the query point, a null result (an unassociated 'res'
pointer) is returned. Bounding the maximum search distance can significantly
reduce the time for queries of points that are far from any point in the tree
(and in which case it is acceptable or desirable to return no result).