Add the Pleim surface layer scheme to MPAS-A
This EPA_PXsfclay commit will add the Pleim surface layer scheme (Pleim,
2006) to MPAS-A as a new "config_sfclayer_scheme" option called
"sf_pxsfclay".  The Pleim surface layer scheme code
(module_sf_pxsfclay.F) was migrated from WRFV4.4.1 with only minor
changes and has been generalized to work with either MPAS or WRF.  Note
that this implementation of the Pleim surface layer in MPAS-A does not
yet support fractional sea ice, in contrast to the other available
surface layer schemes.  The Pleim surface layer works with any LSM or
PBL scheme, but is normally used in conjunction with the PX LSM and the
ACM2 PBL.  Refer to the description in the WRF-ARW technical document
available at https://opensky.ucar.edu/islandora/object/opensky:2898 for
an overview of the physics behind this option.

New file:
  src/core_atmosphere/physics/physics_wrf/module_sf_pxsfclay.F

Modified files:
  Makefile
  src/core_atmosphere/Registry.xml
  src/core_atmosphere/physics/mpas_atmphys_control.F
  src/core_atmosphere/physics/mpas_atmphys_driver_sfclayer.F
  src/core_atmosphere/physics/physics_wrf/Makefile

These EPA_PXsfclay code changes to MPAS-A are based on the 22 August
2022 "develop" branch of MPAS v7.3.

Reference:
Pleim, J. E., 2006: A simple, efficient solution of flux-profile
  relationships in the atmospheric surface layer, J. Appl. Meteor.
  Climatol., 45, 341347, https://doi.org/10.1175/JAM2339.1.