Add MPAS-Atmosphere layer interface projection option
MPAS-Atmosphere vertically projects cell (layer) values (u, rho, theta, etc)
to the level interface (where w is located) using linear interpolation. This
adds the option of using the integrated (average) value between the layers.
This new approach is consistent with the hydrostatic integral used in MPAS-A,
whereas the linear interpolation is not.  This option is set in the
init_atmosphere namelist.

The default setting is still to use linear interpolation.

Change Dynamics 1.3 for Version 8 release.