Add new mpas_halo module to provide group halo communication routines
The new mpas_halo module added in this commit provides the ability to exchange
halos for groups of fields using aggregated messages. The functionality provided
by this new module is similar to what is already available in the mpas_dmpar
module, but the mpas_halo routines are structured in a way that will hopefully
be more amenable to GPU-direct communication of halos with OpenACC. Note that in
this commit, no OpenACC or OpenMP offloading directives are included in the
mpas_halo module.

At present, the new mpas_halo module only supports groups with 2-d and 3-d real-
valued fields, as these two types of fields are sufficient to support all
communication needed in the MPAS-A dynamics, which has served as a use case
during development.

Before halo groups can be constructed and exchanged, the mpas_halo module must
be initilized with a call to mpas_halo_init(...). After all other calls to
routines in the module, a call to mpas_halo_finalize(...) should be made to
enable the module to free resources.

The construction of a halo group is accomplished by a sequence of calls like the
following:

   mpas_halo_exch_group_create

   mpas_halo_exch_group_add_field
   ...
   mpas_halo_exch_group_add_field

   mpas_halo_exch_group_complete

A halo group's resources can be freed by destroying the group with a call to
mpas_halo_exch_group_destroy(...).

After a group has been defined, the halos for all fields in the group can be
exchanged with a call to mpas_halo_exch_group_full_halo_exch(...).