Let config_len_disp default to 0.0, in which case its value is set by nominalMinDc
With this commit, the config_len_disp namelist option defaults to 0.0 and is
hidden by default. Additionally, a new variable, nominalMinDc, is read from the
input file and read/written from/to restart files.

If a positive value for config_len_disp is supplied in the &nhyd_model namelist
group, then that value is used for the horizontal length scale in the model;
otherwise, if a positive value for the nominalMinDc variable is found in the
input file, that value is used.

If both config_len_disp and nominalMinDc values are positive real values, and if
they differ by very much (~1.0e-6 relative difference), then a warning is
printed to the log file, and the value specified in the namelist for
config_len_disp is used to set nominalMinDc.

If neither config_len_disp was specified in namelist.atmosphere file nor a valid
value for nominalMinDc was read from the input file, the atmosphere core
initialization will stop with an error:

  ERROR: Both config_len_disp and nominalMinDc are <= 0.0.
  ERROR: Please either specify config_len_disp in the &nhyd_model namelist group,
  ERROR: or use an input file that provides a valid value for the nominalMinDc variable.