Allow namelist control of coefficient used in Laplacian filter in relaxation zone
The Laplacian filter for momentum in the relaxation zone previously scaled
the divergent component of the Laplacian with a hard-wired coefficient.
With this commit, a new namelist option, config_relax_zone_divdamp_coef, has
been introduced to allow for namelist control over this coefficient.  The
default value of config_relax_zone_divdamp_coef is 6.0, which is different
from the previously hard-wired coefficient of 4.0 in the
atm_bdy_adjust_dynamics_relaxzone_tend routine, reflecting the outcome of
more recent testing.