Always use geog_data_path with trailing slash in mpas_init_atm_static
In cases where the config_geog_data_path namelist option is specified
without a trailing '/' character, trying to build paths to geographical
data tiles at various points in the mpas_init_atm_static module would fail
to produce a valid path. A local variable, geog_data_path, that is the same
as config_geog_data_path but guaranteed to have a trailing '/' was already
being set near the beginning of the init_atm_static() routine, and this
commit simply uses that variable in place of the user-provided
config_geog_data_path.

With the changes in this commit, static field processing can work with a
geographical data path from the namelist that either does or does not have
a trailing '/'.