Allow real-valued attributes to be read with different precision using SMIOL
Previously, the mpas_io_get_att_real0d() routine relied on its 'precision'
optional argument to determine whether to attempt to read real-valued attributes
in a precision different from MPAS's native precision. However, in some parts of
the MPAS framework (specifically, in mpas_bootstrap_framework_phase1()), the
MPAS_io_get_att() routine is called without the 'precision' argument, leading to
problems in reading, e.g., 'sphere_radius' when the precision of the input file
is not the MPAS native precision.

Now, the mpas_io_get_att_real0d() routine first tries to read an attribute in
MPAS's native precision; if that fails, a second attempt is made to read the
attribute using the opposite precision (single vs. double). If both attempts
fail, the attribute argument is set to a fill value and an error code is
returned. This allows MPAS to read real-valued attributes using SMIOL during the
bootstrapping process regardless of the precision of the input file.