Units of adv_coefs?
I've been trying to track units through the dynamics procedure in mpas_atm_time_integration and have found it difficult as several of the units for variables in the Registry are set to unitless when they apparently aren't. 

The units for adv_coefs and adv_coefs_3rd are a great example of ambiguous units, which––assuming that they are unitless––leads to terms in the tend_w calculation that cannot be added owing to differing units. Having dug through the code, it would seem that the units of these terms must be m^3 since they are multiplied by dcEdge**2 and dvEdge in mpas_atm_core.F. However, I am not confident about this, so I've included it as a separate pull request from my other suggested corrections about units.