Fix MPAS field tests
This commit fixes an issue with the field test functions in the test core.
This arose when reading in threadErrs, but not setting it to zero. This
means that the field tests will inheret the error codes from the halo exchange
tests above. intent(out) also allows the compiler to do whatever it desires to
the contents of that array as soon as it enters the function. This could lead
to undefined behavior. The net reseult of this is that the field test
codes are erroneously set to nonzero values while the tests are in fact passing.
After setting all values to 0 before the tests run, the field tests now
correctly show that they are passing.