Add ability to include arbitrary attributes in variables.
This commit adds the functionality needed to add arbitrary attributes to
variables defined in the Registry.xml, and propagate the attributes not
reserved for internal use to the output file. The "reserved" attributes are
contained in the "ATTRS_TO_IGNORE" array, and include name, type, dimensions,
persistence, packages, time_levs, name_in_code, array_group, and default value.
These will continue to be handled as special cases by gen_inc.c. As well, some
attributes will be modified in the code- description will become long_name, and
missing_value becomes _FillValue. A special case for "missing_value" was also
added to the code- there is already logic that determines the proper _FillValue
for a variable, and we continue to use that as the _FillValue when
missing_value is added as an attribute.

In the case of variables defined under a var_array, the missing_value attribute
will be ignored and replaced with the missing_value attribute defined at the
level of the var_array tag. This does not change any behavior from the develop
branch. A new warning is now printed during compilation if a missing_value is
defined in a variable contained within a var_array.

All attributes are now added through the "add_attribute_if_not_ignored" function.
Any changes to the "ATTRS_TO_IGNORE", "ATTRS_TO_MODIFY", or "NUMERIC_ATTRS"
arrays should also include a modification to the "NUM_IGNORED_ATTRS",
"NUM_MODIFIED_ATTRS", or "NUM_NUMERIC_ATTRS" fields, respectively.

Attributes added through the new function will also have single quotes escaped
with the "escape_quotes" function. This simply adds another apostrophe in front
of apostrophes to escape them in fortran.