Add ability to include arbitrary attributes in variables.
This commit adds functionality to include arbitrary variables, along with some
supporting functions. First, it is noted that some attributes do not need to be
added to the the code. These initially include: name, type, dimensions, units,
persistence, packages. As well, the "description" variable must be changed to
the name "long name". Arrays were created to include the ignored and modified
attributes. Sew functions were created to search through these arrays and
either return a signal in the case of the ignored attribute, or replace the
modified attribute. It is noted that these changes initially are not the most
efficient- future commits will likely move the arrays outside of the function
calls.

Another function was created to effectively have the functionality of the old
fortprint calls, but with a bit more clarity and while allowing us to filter
out the ignored or modified attributes. As of the time of this commit,
arbitrary attributes can be added to variables. Future work will include a few
cleanups of some functionality, as well as eliminating redundant code.