Update fortprintf to support Fortran 2003 linebreaks
This commit removes a lot of the logic that was required by the old Fortran 90
linebreak style. This logic was required to know if we were inside of and to
keep track of the indices of those quotes, so that the quotes could be
terminated before a multi line string could be created by concatenating two new
strings together. The Fortran 2003 standard allows for character context
strings to be continued by adding a leading ampersand, such as:

str = 'This is a &
&string'

This allows us to not have to close quotes, and so makes the logic much
simpler. The logic now looks only for being within quotes of any kind. The line
is broken at the closest space to the maximum character limit, and ended with
an ampersand. If we are within quotes and therefore a character context, we lead
the new line with an ampersand. If not, we carry on in the next line as normal.

Two tests were also added, and some incorrect test numbering was
cleaned up in the same file.