Update test Registry.xml to have numeric missing_values
PR #1101 created a new array of names of numeric attributes, one of which is
the "missing_value" attribute. In the test core, however, the only uses of
"missing_value" assign it to the string "FILLVAL". This is nonnumeric, and
causes compiler errors. This commit updates the two uses of "FILLVAL" to
numeric values- 0.0 in the case of the one real valued field, and -999 in the
case of the integer valued field to match other variables in the registry.

The test core compiles and runs properly with these modifications.