Start mpas_string_utils
Throughout the MPAS code, there are a number of utility functions and
subroutines that deal with the manipulation of strings. As more procedures are
added, it begins to make sense to have a common space for these things to be
contained in. To this effect, we have added the "mpas_string_utils" module, and
moved the "mpas_split_string" subroutine from mpas_timekeeping.F to this
module. Future updates will include more functions and subroutines that allow
further manipulation of strings.