Create mpas_string_replace in mpas_string_utils
A function to replace characters in a string with another character has been
added to the mpas_string_utils module. The function searches for a character
(charToReplace) inside of an input string, and replaces any instances of that
character with another character (targetCharacter). Strings are trimmed before
characters are replaced to avoid potential issues with trailing whitespace when
replacing spaces.