Update gen_inc.c to build with default Intel C compiler flags
The Intel C compiler does not default to C99 standards. Introduced in C99 was
the addition of variable declarations within 'for' loops. As a result, when
compiling with the default Intel compiler flags, the variable declaration in
the 'for' loop on line 143 failed to compile. This commit updates the file to
declare the variable the line before, like the other 'for' loops in the file.

The code is now verified to be compliant with the default Intel C standards,
and compiles with the ifort Make target.