Enable supersampling of 30" terrestrial fields: ter, ivgtyp, isltyp, greenfrac
With the introduction of a new namelist option, config_30s_supersample_factor,
in the &data_sources namelist group, the supersampling ratio for several
30-arc-second datasets may now be controlled from the namelist.init_atmosphere
file.

The config_30s_supersample_factor option defaults to 1, which appears to be
reasonable for MPAS meshes with >= 3 km cells. Experimentally, a supersampling
factor of 2 was found to be sufficient for a 1-km quasi-uniform regional mesh
centered over Colorado.

The init_atm_map_static_data routine, which is used by the interp_terrain,
interp_landuse, and interp_soilcat routines, already supported supersampling,
and all that was required for these fields was to pass the namelist
supersampling value down the subroutine call stack.

The code to remap the MODIS FPAR 30-arc-second greenness fraction data did not
support supersampling, and so was modified to use the same approach as for
terrain, land use, and soil category.