Add new MPAS_streamInfo_type derived type and mpas_stream_inquiry module
The new MPAS_streamInfo_type definition has been added in the
mpas_stream_inquiry_types.inc file, and a pointer to an MPAS_streamInfo_type has
been added to the domain_type definition.

The new mpas_stream_inquiry_module provides a routine for returning an instance
of a newly allocated MPAS_streamInfo_type, along with specific routines for
initializing, querying, and finalizing the instance.

The query() function in an MPAS_streamInfo_type accepts the name of a stream,
and optionally, the name of an attribute for that stream. If the stream exists,
and if the optional attribute exists, then the value of the attribute is
returned.

The query() function relies on functions provided in the new stream_inquiry.c
file, which itself makes use of the external ezxml library.

As of this commit, no new code is executed, since the pointer to the
MPAS_streamInfo_type in the domain_type is not set, nor are any procedures
invoked from within an instance.