Instantiate, initialize, and finalize the streamInfo member of the domain
Within the mpas_init routine, after mpas_framework_init_phase2 but
before defining packages, the streamInfo member of the domain is
instantiated and initialized with the MPI communicator and stream
filename for the domain.

In the mpas_finalize routine, after the stream manager has been
finalized but before the log module has been finalized, the streamInfo
member of the domain is finalized and deallocated.