Modifications to clean version 8 of MPAS coe:
1) Code added to ../src/core_atmosphere/mpas_atm_core.F to compute
vertical profiles of epssm that vary with height at the start of the
time integration. These profiles are determined by the parameters
config_min_coeff, config_max_coeff, config_transition_lower_bound,
and config_transition_upper_bound, and are specified in namelist.atmosphere.

2) Code added to allow thermodynamic variables to be represented either
as perurbations from a specified reference state or as the full
variables without a reference state. This option is controlled by the
logical config_reference_sounding as specified in namelist.init_atmosphere.
When configured using the full variables, the horizontal and vertical
pressure gradients computed on the large time steps are written in terms
of log(p).