Merge mpas_dmpar_exch_halo_adj_field2d_real() from the JCSDA-internal/MPAS-Model repository.
Note the number of halo layers impacts the number of cells which will be updated
by this routine:
The first halo layer will update the owned 'edge' cells, where 'edge' cells are
adjacent to ghost cells.
The second halo layer will update owned cells which are adjacent to the 'edge' cells.
The third halo layer will update owned cells which are adjacent to the cells
updated by the seconds halo layer, etc.