Implement new Makefile logic for sharing framework builds across cores
This commit updates the top-level Makefile (and also the Makefile in src/) to
enable reuse of existing builds of either the shared infrastructure or a core
when all build options in the current build are compatible with those used to
previously compile the infrastructure or core.

Previously, after compiling one core but before compiling another, an
intermediate clean step was required (or the AUTOCLEAN=true option could be
specified, which automatically executed the clean step). However, the
infrastructure code in src/framework, src/operators, and src/external is
core-independent and can be reused by multiple MPAS cores, provided build
options are the same.

Now when compiling, build options are saved to the files .build_opts.$(CORE)
and .build_opts.framework.  If either of these files already exist, the build
options saved in them are compared with the current build options to determine
whether the core or the shared infrastructure needs to be recompiled.

The AUTOCLEAN=true option now cleans the core, shared infrastructure, or both as
needed.

Also in this commit, old logic in the top-level Makefile for determining whether
cleaning was necessary has been removed and the .gitignore file has been updated
to ignore the new .build_opts.* files.