Move 21 additional fields into the "invariant" stream in the atmosphere core
This commit moves 21 additional time-invariant fields from the "input" and
"restart" streams into the "invariant" stream for the atmosphere core. These
additional fields are time-invariant terrestrial fields that are used by various
physics parameterization schemes, and as they were in the "input" and "restart"
streams, they are conditionally included in the "invariant" stream with the
DO_PHYSICS preprocessing macro.

Specifically, the 21 fields now included in the "invariant" stream are:

  isltyp
  ivgtyp
  mminlu
  isice_lu
  iswater_lu
  landmask
  shdmin
  shdmax
  snoalb
  albedo12m
  greenfrac
  var2d
  con
  oa1
  oa2
  oa3
  oa4
  ol1
  ol2
  ol3
  ol4