Add OpenACC directives throughout atm_advance_scalars_work
This commit adds OpenACC directives throughout the atm_advance_scalars_work
routine.

Also as part of this commit, macros have been defined at the top of
mpas_atm_time_integration.F to activate timers for OpenACC data transfers when
MPAS-A is compiled with OPENACC=true. For the data transfers in the
atm_advance_scalars_work routine, the new timer

  atm_advance_scalars [ACC_data_xfer]

will be written to the timer summary in log files.

With the NVHPC 24.3 compilers and flags provided in the 'nvhpc' target, results
for a regional test case are bitwise identical when the original code and code
in this commit are both compiled with OPENACC=true.