Add OpenACC directives throughout atm_advance_scalars_mono_work
This commit adds OpenACC directives throughout the atm_advance_scalars_mono_work
routine. At present, halo exchanges in the monotonic scalar transport routine
require fields to be transferred to the host before the halo exchange, and then
back to the device afterward.

For the data transfers in the atm_advance_scalars_mono_work routine, the new
timer

  atm_advance_scalars_mono [ACC_data_xfer]

will be written to the timer summary in log files.

With the NVHPC 24.3 compilers and flags provided in the 'nvhpc' target, results
for a regional test case are bitwise identical when the original code and code
in this commit are both compiled with OPENACC=true.