When running the 'parse' tool, provide Registry file as command-line argument
The 'parse' tool for generating code based on a Registry file can either read
the Registry from stdin or from a file whose name is supplied as a command-line
argument. This tool assumes that if its command-line argument count is 2, then
the argument is a Registry filename (the argument count is 1 if no command-line
arguments are given).

Since future development will add processing for an arbitrary number of
command-line arguments to the parse tool, deciding whether to read the Registry
from stdin based on the argument count is not a viable strategy, so with this
commit, each core's Makefile now passes a filename as the only argument to the
parse tool, and the parse tool itself has been modified to require a
command-line argument.