Add new function 'parse_macros' to registry tool utility code
The parse_macros function takes an array of strings representing C
pre-processing macro definitions (as would be used on the command-line when
invoking cpp or a C/C++ compiler) and parses the macro name and macro value,
passing those as two arguments to a user-provided callback function.

At present, the parse_macros function is not used.