Generate core_variables.inc and domain_variables.inc using run-time macros
Previously, the 'parse' tool used macro definitions at the time it was compiled
to set the content that it writes to the core_variables.inc and
domain_variables.inc files. This has the drawback that the parse tool needs to
be recompiled any time a different core is being compiled.

With this commit, the parse tool now takes additional command-line arguments in
the form of C pre-processor macro definitions to set the values of variables
that are written to core_variables.inc and domain_variables.inc at runtime.

The following macros must now be defined with command-line arguments of the form
-D<macro>=<value> when running the parse tool:

  MPAS_EXE_NAME
  MPAS_GIT_VERSION
  MPAS_BUILD_TARGET
  MPAS_NAMELIST_SUFFIX

For example, when compiling the atmosphere core, the parse tool might be invoked
as

  parse Registry.xml -DMPAS_EXE_NAME=atmosphere_model \
                     -DMPAS_GIT_VERSION=v8.1.0 \
                     -DBUILD_TARGET=gnu \
                     -DNAMELIST_SUFFIX=atmosphere

Accordingly, this commit also modifies the main Makefile for each core so that
the content of $(CPPFLAGS) is passed as command-line arguments to the parse
tool.

Other content written to the core_variables.inc file by the parse tool is still
based on the contents of the Registry.xml file that is supplied as the first
command-line argument.