Parallelize the remapping of Noah-MP static fields in init_atmosphere core
Following the strategy employed for the parallel remapping of other static
fields in the init_atmosphere core, this commit reworks the code for remapping
soilcomp, soilcl1, soilcl2, soilcl3, and soilcl4 so that no special graph
partition files are needed in order to get bit-wise identical values for these
fields when using any number of MPI tasks.

The soilcomp field is handled as a multi-layer, continuous field, while
soilcl{1,2,3,4} are handled as categorical fields (although they are declared
in the Registry as "real" fields).