Add acc parallel and loop directives to port atm_diagnostics_solve_work
These directives are added to perform the do-loop calculations on the
device but more work is required to improve performance.

As with the changes in PR#1176, when using the NVHPC 24.3 compilers with
the 'nvhpc' target and OPENACC=true, these changes do not affect the
results of a regional testcase. Comparisons were made between these
changes and the base branch with all OpenACC directives commented out.