Improve check on deallocation of hash table in mpas_geotile_mgr_finalize
In some cases (typically with the Intel oneAPI compilers), parallel remapping
of static fields in the init_atmosphere core will fail with the message

  ERROR: Problem deallocating the geotile hash table

for some MPI ranks. There is apparently a problem in deallocating the 'hash'
member of mpas_geotile_mgr_type instances in mpas_geotile_mgr_finalize.

This commit improves the checks on the deallocation of 'mgr % hash' in the
mpas_geotile_mgr_finalize routine, making them more stringent. With the
modifications to the deallocation checks, the deallocation errors no longer
occur, suggesting that they were entirely spurious.