Add missing dependency among physics in MPAS-A main physics Makefile
In order to permit correct parallel builds of the MPAS-A physics, the
core_physics target in src/core_atmosphere/physics/Makefile should depend
not only on core_physics_wrf but also on core_physics_noahmp.

Note that core_physics_wrf depends on core_physics_mmm, and so it isn't
necessary to also list core_physics_mmm as a dependency for core_physics.