Add acc data management directives to atm_compute_diagnostics_solve_work
At the beginning of the routine, directives fetch the fields that are
needed on the right-hand side throughout. Then, the fields that are on
the left-hand side are created or copied-in before each parallel region
that needs them. Finally the right-hand side fields are deleted and
the left-hand side fields are copied out at the end of the routine.

Timing for these transfers are reported in the output log file in the
new timer: `atm_compute_solve_diagnostics [ACC_data_xfer]`.

Also, the 'default(present)' clause is added to the parallel directives
so that any missing data on the device causes a runtime error. This is
mainly for development purposes (to ensure the data directives are
correct) and could be removed later on.