Restore CMake support for building the atmosphere core
This commit restores CMake support for the atmosphere core, which broke
due to changes made prior to the version 8.2 release. The breaking
changes included moving the physics_mmm directory to its own git
repository. This is accounted for in CMake by using FetchContent to clone
the correct version of the physics_mmm source code directory based on
the MPAS model git version. Additionally, this commit builds the noahmp
library that was added to the code base. Previously, this library only
supported being built with a hardcoded GNU Makefile. This commit also
accounts for the new registry preprocessing arguments that make the
registry processing core-agnostic. These changes re-enable a seamless
build process for the atmosphere core and improve modularity and
maintainability.