Prepare atm_init_compute_diagnostics for OpenACC porting
In this routine, some integers are pointers but are not arrays. This
causes issues when these variables are used on the GPU with OpenACC.
Variables that are fetched with calls to mpas_atm_get_dimension are
renamed to include '_ptr' and their original names are used for pure
integers.

This also includes using whitespace to align code with the loops so that
OpenACC directives can be easily added.