Update SMIOL code to support specified format for file creation
This commit updates the SMIOL code in src/external/SMIOL from the MPAS-Dev/SMIOL
repository to include modifications that allow the file format to be specified
in calls to SMIOLf_open_file. By default, new output files are still created in
CDF-5 format as before, but it is now possible to specify either
SMIOL_FORMAT_CDF5 or SMIOL_FORMAT_CDF2 for the optional 'fformat' argument to
SMIOLf_open_file.

The changes introduced here come from commit 5df48736 in the MPAS-Dev/SMIOL
repository; but since the SMIOL code in this repository has diverged from that
in the MPAS-Dev/SMIOL repository, the smiol.c file isn't identical between the
two repositories.