Fix MPAS-A dycore-only build
When building MPAS-A as a dycore, all physics-related components are disabled.
An example scenario of this is for use with CAM/CAM-SIMA.

However, the MPAS-A registry file in version 8.2.0 contains the following
regressions, which cause the dycore-only build to break:

* The MPAS-A registry file introduced a new stream, `da_state`, which
unconditionally includes variables that exist only when physics are enabled.
* The MPAS-A registry file unconditionally includes a new registry file from
Noah-MP, which is applicable only when physics are enabled.

Fix MPAS-A dycore-only build by guarding the above regions with the
`DO_PHYSICS` macro.