Add acc data management to atm_set_smlstep_pert_variables_work
Ensure that other, non-invariant, fields are available on the device
during this routine and that modified variables are copied out at the
end.

Timing for these transfers are reported in the output log file in the
new timer: `atm_recover_smlstep_pert_variables [ACC_data_xfer]`.

The 'default(present)' clause is added to the parallel directive so that
any missing data causes a runtime error.