For regional MPAS-A, get the entire 'scalars' state on boundaries at once
Using the implementation of mpas_atm_get_bdy_state for obtaining the boundary
state for 3-d fields, this commit updates the MPAS-A solver to get the state for
all scalars at once, avoiding the need to make one call to the 2-d field
implementation of mpas_atm_get_bdy_state for each scalar.

Not only does the use of the 3-d mpas_atm_get_bdy_state routine simplify the
solver and remove the need for additional calls to the 2-d
mpas_atm_get_bdy_state routine as new scalars are introduced, but the
performance of the dynamics seems to show a slight improvement based on tests
with the GNU 12.2.0 compilers and a six-hour simulation on a regional 12-km mesh
with 156062 cells.