Add halo unit tests for 1-d real fields in test core mpas_halo_tests routine
This commit extends the mpas_halo_tests routine in the test core
mpas_halo_testing module to test the exchange of 1-d real fields using the
mpas_halo module. Now, the exchange groups for persistent and scratch fields
that are tested each include a 1-d real field. With the modifications to the
unit tests, there is no change to the log messages written by the
mpas_halo_testing module.