Introduce new XML attribute, active_when, for <package> elements in the Registry
Prior to the changes in this commit, defining a new package in the Registry with
a <package> element required code to be written (typically in a core's
setup_packages routine) to determine whether that package was active. For simple
conditions -- e.g., a package being active if and only if a logical namelist
option is true -- the need to write boilerplate-type Fortran code to retrieve
package and namelist pointers, check that those pointers are valid, and to
finally set the package variable, was unnecessarily cumbersome.

This commit introduces a new XML attribute, active_when, that may be added to
<package> elements in the Registry, along with new functionality in the registry
processor to automatically generate package-logic code (the equivalent to what
would otherwise need to be written by hand in a core's setup_packages routine)
at compile time.

If provided, the active_when attribute contains a Fortran expression that
evaluates to a logical, and that may use any literal constants and namelist
options (which are assumed to begin with the string 'config'). The expression
from the active_when attribute is essentially used verbatim in Fortran code to
set the package variable.

For example, consider a package named 'foo' that is active if and only if the
namelist option 'config_use_physics' is true. This package might have the
following definition in the Registry:

    <package name="foo" active_when="config_do_physics"/>

Logically, the generated code is equivalent to:

    fooActive = ( config_do_physics )

As another example, consider the following package definition in the Registry:

    <package name="bar" active_when="config_nsteps > 0 .and. .not. config_dryrun"/>

The generated Fortran logic would be:

    barActive = ( config_nsteps > 0 .and. .not. config_dryrun )

When the package's logic involves string literal values, those values must use
single-quote characters in their Registry definition, rather than double-quotes;
for example:

    <package name="blaz" active_when="trim(config_physics_suite) == 'mesoscale'"/>

Note that in these examples, the package's active_when logic is a valid Fortran
logical expression that makes uses of namelist options and literal constants.

In order to make use of this functionality introduced in this commit, a core
must make two changes in its mpas_<CORE>_core_interface module:

1) The module should include through a pre-processor directive the
   registry-generated package-logic code, which resides in a file named
   'setup_packages.inc' in the core's inc/ directory; e.g.,

2) The setup_packages routine in the module should make a call to the generated
   <CORE>_setup_packages_when routine, which is defined in the
   'setup_packages.inc' file; e.g.,

    call atm_setup_packages_when(configs, packages)