Add init_atm function to check if grid is a unit sphere
This function uses the sum of cellArea for all owned cells to
approximate the surface area of the sphere. This function returns true
if the sum of cellArea is close to unit sphere surface area. Otherwise
this function returns false.

cellAreas are simple to sum, while direct checks of distance of cell
centers to the origin could be more expensive.