Add build check for PnetCDF to top-level Makefile
Adds a build target (pnetcdf_test) that attempts to build a test PnetCDF
C program. Failing to build the test program causes a warning message to
the user, leaves the test program and build log in the top-level
directory, and aborts the MPAS build.

The MPAS build will also abort if the pnetcdf_test detects the PNETCDF
environment variable isn't set.