Add acc data movement to atm_bdy_adjust_scalars_work
Ensure that the other, non-invariant fields are available for this
routine. Variables that are overwritten during this routine are only
created while others are copied in. Any variables overwritten by this
routine are copied out at the end.

Timing for these transfers are reported in the output log file in the
new timer: 'atm_bdy_adjust_scalars [ACC_data_xfer]'.

Also add default(present) to parallel directives to ensure data movement
is correct and prevent any implicit data movements from the compiler.