Add MPAS wrapper module in the physics_wrf directory for the UGWP schemes
This commit adds a new module, module_bl_ugwp_gwdo, in the
core_atmosphere/physics/physics_wrf directory. This module serves as a wrapper
for the UGWP physics that is compatible with the MPAS-A physics drivers.

Changes to the Makefile in the physics_wrf directory ensure that the
module_bl_ugwp_gwdo.F file is compiled, but at present the code in that module
is not actually called.