Make SMIOL the default I/O layer when running cmake.
The default I/O layer when building MPAS-Model via cmake used to be PIO,
but when building via make SMIOL is the default I/O layer.
This change makes SMIOL the default I/O layer when building via cmake.

The cmake variable MPAS_USE_SMIOL is changed to be MPAS_USE_PIO.
The default value of MPAS_USE_PIO is OFF.
To build with the PIO library run cmake with -DMPAS_USE_PIO=ON.