Add a simple test definition for testing GNU 12.2.0 compilation
Following the documentation of the hpc-workflows testing framework and
the testing structure found in .ci/, a JSON file for a GNU compilation
test was added.

This test will compile the atmosphere core using gnu + single precision.
If this test is run using the derecho configuration the appropriate
modules will attempt to be loaded. For non-derecho environments, per the
testing structure under .ci/, if no configuration exists in
.ci/hostenv.sh then the current environment will be used verbatim.