Initial OpenACC port of mpas_vector_reconstruct_2d
Ensures the data needed for the mpas_reconstruct_2d routine has
been fetched onto the device (GPU) at the beginning and end of the
routine. The time for these transfers are captured in a new timer
'mpas_reconstruct_2d [ACC_data_xfer]'. This is enforced by the
default(present) clauses.

NOTE: coeffs_reconstruct, nEdgesOnCell, edgesOnCell, latCell, and
lonCell are also fetched in mpas_reconstruct_2d. This is because this
routine is called before these variables would be uploaded to the device
during mpas_atm_dynamics_init as part of atmosphere_core initialization.
The copyins will not execute anymore once the model starts timestepping
and the OpenACC runtime sees the variables are present on the device.