Use 2 loop search for min/max with location in summarize_timestep
Only for loops executed when 'config_print_detailed_minmax_vel = .true.'

When running with OpenACC, it is difficult to get values to "come with"
during a reduction. This means it's difficult to get both the minimum
value of an array and the position (the array indices) it occurs at in
one pass.  Using 2 loops like this ensures the correct positions are
found.

Using a "linear index" to record the position of the minimum/maximum
values helps in cases where the min/max occurs at multiple points in an
array. The detailed prints will always use the first occurence of the
min/max in the array.