Add support for linking with the MUSICA-Fortran library
This commit adds support for linking with the MUSICA-Fortran library. This
capability is intended as a first step towards being able to use MUSICA
chemistry components (MICM, TUV-x) in MPAS-Atmosphere.

In order for MPAS to link with the MUSICA-Fortran library, the library must
be detectable through 'pkg-config'. Specifying MUSICA=true on the MPAS 'make'
command line will cause MPAS to link with the MUSICA library; for example:

  make gnu CORE=atmosphere MUSICA=true

Note that, although the use of the MUSICA library will likely be within the
context of the atmosphere core only, compiling any core with MUSICA=true will
cause the MUSICA library to be linked with the MPAS model core executable.

If the MUSICA-Fortran library has been linked successfully, the build summary
will include the message

  MPAS was linked with the MUSICA-Fortran library version 0.10.1.

(possibly with a different version number); additionally, when running the
model, the log file will contain a message indicating MUSICA support as well
as the MICM library version, e.g.:

   MUSICA support: yes
   - MICM version: 3.8.0

If the MUSICA-Fortran library has not been linked, the build summary will
include the message

  MPAS was not linked with the MUSICA-Fortran library.

and the log file for a core will contain the message

   MUSICA support: no

.