Convert mpas_atm_get_bdy_tend from a function to a subroutine
The mpas_atm_get_bdy_tend function has been converted to a subroutine, with the
field that it previously returned now provided to the caller as an intent(out)
argument. Previous invocations of the mpas_atm_get_bdy_tend function in atm_srk3
have been updated to call the mpas_atm_get_bdy_tend subroutine.

The change to a subroutine is driven by the desire to keep the entire dynamics
state resident in device memory when MPAS-A is being run on GPUs with OpenACC:
a subroutine argument can be copied to the device before the call to the
subroutine, and the subroutine can then directly work on the device to modify
that argument; in contrast, returning an on-device array from a function is
less straightforward with OpenACC. As of this commit, no changes to data
movement have been made, however.